+++ Source/Client/ArbiterPatches.cs
@@ -1,12 +1,10 @@
 ﻿using Harmony;
 using RimWorld;
 using RimWorld.Planet;
-using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Reflection;
 using System.Reflection.Emit;
-using System.Text;
 using UnityEngine;
 using Verse;
