+++ Source/Client/Blueprints.cs
@@ -2,10 +2,8 @@
 using RimWorld;
 using System;
 using System.Collections.Generic;
-using System.Linq;
 using System.Reflection;
 using System.Reflection.Emit;
-using System.Text;
 using Verse;
 
 namespace Multiplayer.Client
