+++ Source/Client/CacheForReloading.cs
@@ -3,7 +3,6 @@ using RimWorld.Planet;
 using System;
 using System.Collections.Generic;
 using System.Linq;
-using System.Text;
 using Verse;
 
 namespace Multiplayer.Client
