+++ Source/Client/ClientConnection.cs
@@ -1,10 +1,6 @@
 ﻿using System;
-using System.Collections;
 using System.Collections.Generic;
-using System.Diagnostics;
 using System.Linq;
-using System.Security.Cryptography;
-using System.Text;
 using Verse;
 
 namespace Multiplayer.Client
