+++ Source/Client/CrossRefs.cs
@@ -2,7 +2,6 @@
 using RimWorld;
 using RimWorld.Planet;
 using Verse;
-using Verse.Profile;
 
 namespace Multiplayer.Client
 {
