+++ Source/Client/Designators.cs
@@ -1,11 +1,8 @@
 ﻿using Harmony;
 using Multiplayer.Common;
 using RimWorld;
-using System;
 using System.Collections.Generic;
 using System.Linq;
-using System.Reflection;
-using System.Reflection.Emit;
 using Multiplayer.Client.Comp;
 using Multiplayer.Client.Sync;
 using Multiplayer.Common.Networking;
