+++ Source/Client/MpPatch.cs
@@ -5,7 +5,6 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Reflection;
 using System.Reflection.Emit;
-using Verse;
 
 namespace Multiplayer.Client
 {
