+++ Source/Client/MpReflection.cs
@@ -1,7 +1,5 @@
 ﻿using Harmony;
-using Multiplayer.Common;
 using System;
-using System.Collections;
 using System.Collections.Generic;
 using System.Linq;
 using System.Reflection;
