+++ Source/Client/MpUtil.cs
@@ -1,11 +1,8 @@
 ﻿using Harmony;
 using Harmony.ILCopying;
-using Multiplayer.Common;
-using Steamworks;
 using System;
 using System.Collections;
 using System.Collections.Generic;
-using System.Diagnostics;
 using System.Linq;
 using System.Net;
 using System.Net.Sockets;
@@ -14,8 +11,6 @@ using System.Reflection.Emit;
 using System.Runtime.CompilerServices;
 using System.Runtime.InteropServices;
 using System.Runtime.Serialization;
-using System.Text;
-using System.Threading;
 using UnityEngine;
 using Verse;
