+++ Source/Client/MultiplayerMod.cs
@@ -1,21 +1,13 @@
 ﻿using Harmony;
-using Harmony.ILCopying;
 using Multiplayer.Common;
 using RimWorld;
 using System;
-using System.Collections;
 using System.Collections.Generic;
-using System.Diagnostics;
-using System.Globalization;
-using System.Linq;
 using System.Reflection;
 using System.Reflection.Emit;
-using System.Runtime.CompilerServices;
-using System.Runtime.InteropServices;
-using System.Text;
 using System.Xml;
 using Multiplayer.Server;
-using Multiplayer.Server.Networking.State;
+using Multiplayer.Server.Networking.Handler;
 using UnityEngine;
 using Verse;
