+++ Source/Client/MultiplayerSession.cs
@@ -3,12 +3,10 @@ using Multiplayer.Common;
 using RimWorld;
 using RimWorld.BaseGen;
 using Steamworks;
-using System;
 using System.Collections.Generic;
 using System.Diagnostics;
 using System.Linq;
 using System.Reflection;
-using System.Text;
 using System.Threading;
 using Multiplayer.Client.Comp;
 using Multiplayer.Client.Desyncs;
