+++ Source/Client/Networking/SteamClientToServerConnection.cs
@@ -1,4 +1,3 @@
-using Multiplayer.Common;
 using Multiplayer.Common.Networking;
 using Multiplayer.Common.Networking.Connection;
 using Steamworks;
