+++ Source/Client/OnMainThread.cs
@@ -1,15 +1,12 @@
 ﻿using Multiplayer.Common;
 using RimWorld.Planet;
-using Steamworks;
 using System;
 using System.Collections.Generic;
-using System.Diagnostics;
 using System.Linq;
 using Multiplayer.Client.Sync;
 using Multiplayer.Client.Windows;
 using Multiplayer.Common.Networking;
 using Multiplayer.Common.Networking.Connection;
-using Multiplayer.Server;
 using UnityEngine;
 using Verse;
 using Verse.Steam;
