+++ Source/Client/Optimizations.cs
@@ -1,5 +1,4 @@
 ﻿using Harmony;
-using Multiplayer.Common;
 using System;
 using System.Collections;
 using System.Collections.Generic;
