+++ Source/Client/Patches.cs
@@ -1,18 +1,14 @@
 ﻿using Harmony;
-using Harmony.ILCopying;
 using Multiplayer.Common;
 using RimWorld;
 using RimWorld.Planet;
-using Steamworks;
 using System;
 using System.Collections.Generic;
-using System.Diagnostics;
 using System.Linq;
 using System.Reflection;
 using System.Reflection.Emit;
 using System.Runtime.CompilerServices;
 using System.Text.RegularExpressions;
-using System.Xml;
 using System.Xml.Linq;
 using Multiplayer.Client.Comp;
 using Multiplayer.Client.Windows;
