+++ Source/Client/PlayerCursors.cs
@@ -1,10 +1,8 @@
 ﻿using Harmony;
 using Multiplayer.Common;
 using RimWorld;
-using System;
 using System.Collections.Generic;
 using System.Linq;
-using System.Text;
 using UnityEngine;
 using Verse;
