+++ Source/Client/PlayerListEntry.cs
@@ -1,5 +1,4 @@
 using System.Collections.Generic;
-using Multiplayer.Common;
 using Multiplayer.Common.Networking;
 using Multiplayer.Server;
 using UnityEngine;
