+++ Source/Client/SaveCompression.cs
@@ -2,11 +2,7 @@
 using RimWorld;
 using System;
 using System.Collections.Generic;
-using System.Diagnostics;
 using System.IO;
-using System.Linq;
-using System.Reflection;
-using UnityEngine;
 using Verse;
 using Verse.Profile;
