+++ Source/Client/SaveLoad.cs
@@ -3,11 +3,9 @@ using Ionic.Zlib;
 using Multiplayer.Common;
 using RimWorld;
 using RimWorld.Planet;
-using System;
 using System.Collections.Generic;
 using System.Diagnostics;
 using System.Linq;
-using System.Text;
 using System.Threading;
 using System.Xml;
 using Multiplayer.Client.Comp;
