+++ Source/Client/ScribeUtil.cs
@@ -1,5 +1,4 @@
 ﻿using Harmony;
-using Multiplayer.Common;
 using RimWorld;
 using System;
 using System.Collections.Generic;
