+++ Source/Client/Seeds.cs
@@ -1,14 +1,10 @@
 ﻿using Harmony;
-using Multiplayer.Common;
 using RimWorld;
 using RimWorld.Planet;
 using System;
-using System.Collections;
 using System.Collections.Generic;
-using System.Linq;
 using System.Reflection;
 using System.Reflection.Emit;
-using UnityEngine;
 using Verse;
 using Verse.Grammar;
