+++ Source/Client/SetMapTime.cs
@@ -2,9 +2,6 @@
 using RimWorld;
 using RimWorld.Planet;
 using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
 using Verse;
 using Verse.Sound;
