+++ Source/Client/SettingsPatches.cs
@@ -1,10 +1,5 @@
 ﻿using Harmony;
 using RimWorld;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using UnityEngine;
 using Verse;
 
 namespace Multiplayer.Client
