+++ Source/Client/SituationalThoughts.cs
@@ -1,9 +1,6 @@
 ﻿using Harmony;
 using RimWorld;
-using System;
 using System.Collections.Generic;
-using System.Linq;
-using System.Text;
 using Verse;
 
 namespace Multiplayer.Client
