+++ Source/Client/TickPatch.cs
@@ -7,7 +7,6 @@ using System;
 using System.Collections.Generic;
 using System.Diagnostics;
 using System.Linq;
-using System.Threading;
 using Multiplayer.Client.Comp;
 using UnityEngine;
 using Verse;
