+++ Source/Client/TileTemperatures.cs
@@ -1,9 +1,5 @@
 ﻿using Harmony;
 using RimWorld.Planet;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
 using Verse;
 
 namespace Multiplayer.Client
