+++ Source/Client/Windows/ChatWindow.cs
@@ -4,7 +4,7 @@ using System.Text;
 using Multiplayer.Common;
 using Multiplayer.Common.Networking;
 using Multiplayer.Server;
-using Multiplayer.Server.Networking.State;
+using Multiplayer.Server.Networking.Handler;
 using RimWorld;
 using Steamworks;
 using UnityEngine;
