+++ Source/Client/Windows/ServerBrowser.cs
@@ -12,7 +12,6 @@ using LiteNetLib;
 using Multiplayer.Client.Desyncs;
 using Multiplayer.Client.Networking;
 using Multiplayer.Common;
-using Multiplayer.Common.Networking;
 using Multiplayer.Common.Networking.Handler;
 using Multiplayer.Server;
 using RimWorld;
