+++ Source/Common/ChatCommands.cs
@@ -1,8 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using Multiplayer.Common.Networking;
+﻿using Multiplayer.Common.Networking;
 using Multiplayer.Server;
 
 namespace Multiplayer.Common
