+++ Source/Common/Extensions.cs
@@ -8,7 +8,6 @@ using System.Reflection;
 using System.Security.Cryptography;
 using System.Text;
 using System.Xml;
-using Multiplayer.Common.Networking;
 using Multiplayer.Common.Networking.Connection;
 
 namespace Multiplayer.Common
