+++ Source/Server/Networking/SteamServerToClientConnection.cs
@@ -1,5 +1,4 @@
-﻿using Multiplayer.Common;
-using Multiplayer.Common.Networking;
+﻿using Multiplayer.Common.Networking;
 using Multiplayer.Common.Networking.Connection;
 using Steamworks;
